Move files into circular_slider folder
