+++ lib/music_player/pitch_speed_card/circular_slider/circular_slider.dart
@@ -3,8 +3,8 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:musbx/music_player/pitch_speed_card/circular_slider/painter.dart';
 import 'package:musbx/music_player/pitch_speed_card/circular_slider/utils.dart';
-import 'package:musbx/music_player/pitch_speed_card/circular_slider_theme.dart';
-import 'package:musbx/music_player/pitch_speed_card/custom_pan_gesture_recognizer.dart';
+import 'package:musbx/music_player/pitch_speed_card/circular_slider/theme.dart';
+import 'package:musbx/music_player/pitch_speed_card/circular_slider/custom_pan_gesture_recognizer.dart';
 
 enum DraggingMode {
   none,
