+++ lib/music_player/pitch_speed_card/circular_slider/custom_pan_gesture_recognizer.dart
