+++ lib/music_player/pitch_speed_card/circular_slider/theme.dart
